# src/main/java/ru/neoflex/openapi/controllers/ApplicationApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package ru.neoflex.openapi.controllers;

import ru.neoflex.openapi.dtos.ErrorResponse;
import ru.neoflex.openapi.dtos.LoanApplicationRequest;
import ru.neoflex.openapi.dtos.LoanOffer;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-08-29T12:43:58.086982+03:00[Europe/Moscow]")
@Validated
@Tag(name = "application", description = "the application API")
public interface ApplicationApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * PUT /application/offer : Choosing one of the offers
     *
     * @param loanOffer  (required)
     * @return Offer applied (status code 200)
     *         or Bad request (status code 400)
     */
    @Operation(
        operationId = "applyOffer",
        summary = "Choosing one of the offers",
        responses = {
            @ApiResponse(responseCode = "200", description = "Offer applied"),
            @ApiResponse(responseCode = "400", description = "Bad request", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ErrorResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/application/offer",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> _applyOffer(
        @Parameter(name = "LoanOffer", description = "", required = true) @Valid @RequestBody LoanOffer loanOffer
    ) {
        return applyOffer(loanOffer);
    }

    // Override this method
    default  ResponseEntity<Void> applyOffer(LoanOffer loanOffer) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /application : Calculation of loan terms
     *
     * @param loanApplicationRequest  (required)
     * @return Application created (status code 200)
     *         or Bad request (status code 400)
     */
    @Operation(
        operationId = "createApplication",
        summary = "Calculation of loan terms",
        responses = {
            @ApiResponse(responseCode = "200", description = "Application created", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = LoanOffer.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ErrorResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/application",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<List<LoanOffer>> _createApplication(
        @Parameter(name = "LoanApplicationRequest", description = "", required = true) @Valid @RequestBody LoanApplicationRequest loanApplicationRequest
    ) {
        return createApplication(loanApplicationRequest);
    }

    // Override this method
    default  ResponseEntity<List<LoanOffer>> createApplication(LoanApplicationRequest loanApplicationRequest) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"totalAmount\" : 10150.0, \"isSalaryClient\" : true, \"monthlyPayment\" : 1756.38, \"rate\" : 13, \"requestedAmount\" : 10000.0, \"term\" : 6, \"applicationId\" : 1, \"isInsuranceEnabled\" : true }, { \"totalAmount\" : 10150.0, \"isSalaryClient\" : true, \"monthlyPayment\" : 1756.38, \"rate\" : 13, \"requestedAmount\" : 10000.0, \"term\" : 6, \"applicationId\" : 1, \"isInsuranceEnabled\" : true }, { \"totalAmount\" : 10150.0, \"isSalaryClient\" : true, \"monthlyPayment\" : 1756.38, \"rate\" : 13, \"requestedAmount\" : 10000.0, \"term\" : 6, \"applicationId\" : 1, \"isInsuranceEnabled\" : true }, { \"totalAmount\" : 10150.0, \"isSalaryClient\" : true, \"monthlyPayment\" : 1756.38, \"rate\" : 13, \"requestedAmount\" : 10000.0, \"term\" : 6, \"applicationId\" : 1, \"isInsuranceEnabled\" : true } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}